changed project structure, divided into modules, added repositories and facades for all modules
