+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/carton/Carton.java
@@ -1,19 +1,20 @@
-package pl.pkasiewicz.filterpackmate.domain.entity;
+package pl.pkasiewicz.filterpackmate.domain.carton;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import pl.pkasiewicz.filterpackmate.domain.product.Product;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Data
-public class Side {
+public class Carton {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
     String name;
-    @ManyToMany(mappedBy = "sides")
+    @OneToMany(mappedBy = "carton")
     List<Product> products = new ArrayList<>();
 }
