+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/carton/CartonFacade.java
@@ -1,0 +1,4 @@
+package pl.pkasiewicz.filterpackmate.domain.carton;
+
+public class CartonFacade {
+}
