+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/carton/CartonRepository.java
@@ -1,0 +1,8 @@
+package pl.pkasiewicz.filterpackmate.domain.carton;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface CartonRepository extends JpaRepository<Carton, Long> {
+}
