+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/divider/Divider.java
@@ -1,19 +1,20 @@
-package pl.pkasiewicz.filterpackmate.domain.entity;
+package pl.pkasiewicz.filterpackmate.domain.divider;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import pl.pkasiewicz.filterpackmate.domain.product.Product;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Data
-public class Tray {
+public class Divider {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
     String name;
-    @OneToMany(mappedBy = "tray")
+    @ManyToMany(mappedBy = "dividers")
     List<Product> products = new ArrayList<>();
 }
