+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/divider/DividerFacade.java
@@ -1,0 +1,5 @@
+package pl.pkasiewicz.filterpackmate.domain.divider;
+
+public class DividerFacade {
+
+}
