+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/divider/DividerRepository.java
@@ -1,0 +1,8 @@
+package pl.pkasiewicz.filterpackmate.domain.divider;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DividerRepository extends JpaRepository<Divider, Long> {
+}
