--- src/main/java/pl/pkasiewicz/filterpackmate/domain/entity/Pallet.java
@@ -1,5 +1,0 @@
-package pl.pkasiewicz.filterpackmate.domain.entity;
-
-enum Pallet {
-    EURO, GM6, HG5, J7
-}
