+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/product/Pallet.java
@@ -1,0 +1,5 @@
+package pl.pkasiewicz.filterpackmate.domain.product;
+
+public enum Pallet {
+    EURO, GM6, HG5, J7
+}
