+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/product/ProductFacade.java
@@ -1,0 +1,4 @@
+package pl.pkasiewicz.filterpackmate.domain.product;
+
+public class ProductFacade {
+}
