+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/side/Side.java
@@ -1,19 +1,20 @@
-package pl.pkasiewicz.filterpackmate.domain.entity;
+package pl.pkasiewicz.filterpackmate.domain.side;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import pl.pkasiewicz.filterpackmate.domain.product.Product;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Data
-public class Divider {
+public class Side {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
     String name;
-    @ManyToMany(mappedBy = "dividers")
+    @ManyToMany(mappedBy = "sides")
     List<Product> products = new ArrayList<>();
 }
