+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/side/SideFacade.java
@@ -1,0 +1,4 @@
+package pl.pkasiewicz.filterpackmate.domain.side;
+
+public class SideFacade {
+}
