+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/side/SideRepository.java
@@ -1,0 +1,8 @@
+package pl.pkasiewicz.filterpackmate.domain.side;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface SideRepository extends JpaRepository<SideRepository, Long> {
+}
