+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/tray/TrayFacade.java
@@ -1,0 +1,4 @@
+package pl.pkasiewicz.filterpackmate.domain.tray;
+
+public class TrayFacade {
+}
