+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/tray/TrayRepository.java
@@ -1,0 +1,8 @@
+package pl.pkasiewicz.filterpackmate.domain.tray;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface TrayRepository  extends JpaRepository<Tray, Long> {
+}
