+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/user/User.java
@@ -1,4 +1,4 @@
-package pl.pkasiewicz.filterpackmate.domain.entity;
+package pl.pkasiewicz.filterpackmate.domain.user;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -8,7 +8,7 @@ import lombok.Data;
 
 @Entity
 @Data
-public class User {
+class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
