+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/user/UserFacade.java
@@ -1,0 +1,5 @@
+package pl.pkasiewicz.filterpackmate.domain.user;
+
+public class UserFacade {
+
+}
