+++ src/main/java/pl/pkasiewicz/filterpackmate/domain/user/UserRepository.java
@@ -1,0 +1,9 @@
+package pl.pkasiewicz.filterpackmate.domain.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface UserRepository extends JpaRepository<User, Long> {
+
+}
